+++ bundles/framework/bundle/featuredata2/plugin/FeaturedataPlugin.js
@@ -153,7 +153,7 @@ function(config) {
         var me = this;
 
         // get div where the map is rendered from openlayers
-        var parentContainer = me.mapModule.getMapLayersContainerDiv();
+        var parentContainer = jQuery(this._map.div);
 
         if(!me.__elements['main']) {
             me.__elements['main'] = me.__templates['main'].clone();
@@ -167,7 +167,7 @@ function(config) {
         me.__elements['main'].mousedown(function(event) {
             event.stopPropagation();
         });
-        parentContainer.appendChild(me.__elements['main'][0]);        
+        parentContainer.append(me.__elements['main']);        
         this.update();
     },