+++ packages/statistics/bundle/publishedgrid/bundle.js
@@ -33,6 +33,9 @@ function() {
             "src" : "../../../../bundles/statistics/bundle/statsgrid/plugin/ManageStatsPlugin.js"
         }, {
             "type" : "text/javascript",
+            "src" : "../../../../bundles/statistics/bundle/statsgrid/event/SotkadataChangedEvent.js"
+        }, {
+            "type" : "text/javascript",
             "src" : "../../../../bundles/statistics/bundle/statsgrid/service/StatisticsService.js"
         }, {
             "type" : "text/css",