+++ tests/statistics/statsgrid/statisticsService.js
@@ -1,4 +1,4 @@
-describe.only('Test Suite for statistics/statsgrid statistics service', function() {
+describe('Test Suite for statistics/statsgrid statistics service', function() {
     var appSetup = null,
         appConf = null,
         statsModule = null,